+++ examples/jsm/nodes/code/FunctionCallNode.js
@@ -1,5 +1,5 @@
-import TempNode from './TempNode.js';
-import { addNodeClass } from './Node.js';
+import TempNode from '../core/TempNode.js';
+import { addNodeClass } from '../core/Node.js';
 import { addNodeElement, nodeArray, nodeObject, nodeObjects } from '../shadernode/ShaderNode.js';
 
 class FunctionCallNode extends TempNode {
