+++ examples/jsm/nodes/code/ScriptableNode.js
