+++ examples/jsm/nodes/code/ScriptableValueNode.js
