+++ examples/jsm/nodes/materialx/lib/mx_hsv.js
@@ -1,4 +1,4 @@
-import { fn } from '../../core/FunctionNode.js';
+import { fn } from '../../code/FunctionNode.js';
 
 // Original shader code from:
 // https://github.com/AcademySoftwareFoundation/MaterialX/blob/main/libraries/stdlib/genglsl/lib/mx_hsv.glsl
