+++ examples/jsm/nodes/materialx/lib/mx_noise.js
@@ -1,5 +1,5 @@
-import { code } from '../../core/CodeNode.js';
-import { fn } from '../../core/FunctionNode.js';
+import { code } from '../../code/CodeNode.js';
+import { fn } from '../../code/FunctionNode.js';
 
 // Original shader code from:
 // https://github.com/AcademySoftwareFoundation/MaterialX/blob/main/libraries/stdlib/genglsl/lib/mx_noise.glsl
